client cant upload the same file twice (naive cause its based on the filename but hey

+++ back/systemyRozproszone/src/main/java/pl/systemyRozproszone/systemyRozproszone/fileUpload/FileUploadController.java
@@ -16,18 +16,47 @@ import java.io.IOException;
 @RestController
 public class FileUploadController {
 
+    /**
+     *
+     * @param file - file from the request
+     * @return http status
+     *      if file exists in given directory it returns 406 NOT_ACCEPTABLE
+     *      if file doesnt exist - it returns 200 OK
+     */
     @PostMapping(value = "/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<Object> uploadFile(@RequestParam("file") MultipartFile file){
-        File newFile = new File("/Users/tomaszkoltun/Documents/uploadedSpringFiles/"+file.getOriginalFilename());
-        try {
-            newFile.createNewFile();
-            FileOutputStream fout = new FileOutputStream(newFile);
-            fout.write(file.getBytes());
-            fout.close();
-        } catch (IOException e) {
-            e.printStackTrace();
+        
+        boolean doesFileAlreadyExist = checkWhetherFileWithThisNameAlreadyExists(file.getOriginalFilename());
+
+        if(!doesFileAlreadyExist){
+            File newFile = new File("${fileUploadDirectory}"+file.getOriginalFilename());
+            try {
+                newFile.createNewFile();
+                FileOutputStream fout = new FileOutputStream(newFile);
+                fout.write(file.getBytes());
+                fout.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            return new ResponseEntity<>("file uploaded", HttpStatus.OK);
         }
 
-        return new ResponseEntity<>("file uploaded", HttpStatus.OK);
+        return new ResponseEntity<>("file already exists", HttpStatus.NOT_ACCEPTABLE);
+    }
+
+    private boolean checkWhetherFileWithThisNameAlreadyExists(String fileName) {
+        File dir = new File("${fileUploadDirectory}");
+        File[] directoryListing = dir.listFiles();
+        if (directoryListing != null) {
+            for (File child : directoryListing) {
+                if(child.getName().equals(fileName)){
+                    return true;
+                }
+            }
+        } else {
+            return true;
+        }
+        return false;
     }
 }
